feat(file_ops): complete Phase 4 - remove embedded file_ops component
Removes the deprecated embedded Go-based file operations component and finalizes
the migration to the WASM-based component architecture.

Changes:
- Delete tools/file_ops/ embedded Go binary
- Rename tools/file_ops_external → tools/file_ops (single implementation now)
- Add wit_files filegroup to tools/file_ops/BUILD.bazel
- Move wit/file-operations.wit from embedded to external location
- Update all references from file_ops_external to file_ops
- Update toolchains/BUILD.bazel to use file_ops with wit_files
- Update test/file_ops_integration/file_ops_test.bzl to use file_ops
- Remove deprecated file_ops_source flag (embedded vs external choice)
- Update MODULE.bazel toolchain registration
- Add platform_name field to wac.json for registry consistency

This completes Phase 4 migration. The single file_ops implementation now uses
the external WASM component under the hood, providing:
- 100x faster startup with AOT compilation support
- Improved security through WASM sandboxing
- Cross-platform compatibility

Fixes #183
