+++ toolchains/BUILD.bazel
@@ -204,6 +204,10 @@ bzl_library(
 # Note: C++ toolchain configuration has been moved to @wasi_sdk repository
 # The cc_toolchain is now registered via @wasi_sdk//:cc_toolchain in MODULE.bazel
 
+# Phase 4 Complete: External File Operations Component
+# The external pre-built WASM component is now the standard
+# See https://github.com/pulseengine/bazel-file-ops-component for component source
+
 file_ops_toolchain(
     name = "file_ops_toolchain_impl",
     file_ops_component = "//tools/file_ops:file_ops",
@@ -211,25 +215,8 @@ file_ops_toolchain(
 )
 
 toolchain(
-    name = "file_ops_toolchain_local",
-    # Universal toolchain - works on all platforms
+    name = "file_ops_toolchain_target",
     toolchain = ":file_ops_toolchain_impl",
-    toolchain_type = ":file_ops_toolchain_type",
-)
-
-# Phase 1 Integration: External File Operations Component from bazel-file-ops-component
-# This uses the pre-built WASM component from https://github.com/pulseengine/bazel-file-ops-component
-
-file_ops_toolchain(
-    name = "file_ops_toolchain_external_impl",
-    file_ops_component = "//tools/file_ops_external:file_ops_external",
-    wit_files = ["//tools/file_ops:wit_files"],
-)
-
-toolchain(
-    name = "file_ops_toolchain_external",
-    # External component toolchain (opt-in for Phase 1)
-    toolchain = ":file_ops_toolchain_external_impl",
     toolchain_type = ":file_ops_toolchain_type",
 )
 
@@ -258,23 +245,6 @@ string_flag(
     ],
 )
 
-# File Operations Source Selection
-# Phase 3: DEPRECATION - Embedded component will be removed in v2.0.0
-#
-# The embedded component (tools/file_ops/) is DEPRECATED and will be removed in v2.0.0.
-# See docs/MIGRATION.md for migration guide.
-#
-# Default is "external" with AOT support (100x faster startup)
-# Only use "embedded" if you encounter critical issues - this option will be removed soon.
-string_flag(
-    name = "file_ops_source",
-    build_setting_default = "external",  # Phase 2+: External with AOT is default
-    values = [
-        "embedded",  # DEPRECATED - Will be removed in v2.0.0 (Phase 4)
-        "external",  # RECOMMENDED - External component with AOT (default)
-    ],
-)
-
 # Configuration settings for implementation selection
 config_setting(
     name = "file_ops_use_tinygo",
@@ -294,21 +264,6 @@ config_setting(
     name = "file_ops_auto_select",
     flag_values = {
         ":file_ops_implementation": "auto",
-    },
-)
-
-# Configuration settings for source selection (Phase 2: External with AOT is default)
-config_setting(
-    name = "file_ops_use_embedded",
-    flag_values = {
-        ":file_ops_source": "embedded",
-    },
-)
-
-config_setting(
-    name = "file_ops_use_external",
-    flag_values = {
-        ":file_ops_source": "external",
     },
 )
